# backend/progresstracker/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Per-field validation messages. Present on validation failures only.",
            "type" : "object"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        }
      },
      "ProfileRequest" : {
        "properties" : {
          "timeZone" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [
          "timeZone"
        ],
        "type" : "object"
      },
      "ProfileResponse" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "timeZone" : {
            "description" : "IANA time zone; the user's days are counted in it",
            "type" : "string"
          }
        },
        "required" : [
          "email",
          "timeZone"
        ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Habits, completions, streaks, XP, and achievements.",
    "title" : "Progress Tracker API",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/me" : {
      "get" : {
        "operationId" : "getProfile",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          }
        },
        "tags" : [
          "Profile"
        ]
      },
      "put" : {
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request body is malformed or fails validation"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid bearer token"
          }
        },
        "tags" : [
          "Profile"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearerAuth" : [ ]
    }
  ]
}